+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 * Allowed `TxControl` in `database/sql` transaction execution when TxControl equals TxControl on BeginTx
+
+## v3.126.0
+* Added `query.WithLazyTx(bool)` option for `query.Client.DoTx` calls to enable/disable lazy transactions per operation
+* Added `retry.WithLazyTx(bool)` option for `retry.DoTx` calls to enable/disable lazy transactions for database/sql
+
+## v3.125.4
 * Added support for `GlobalUniqueIndex` type in `table/options.IndexDescription`
 
 ## v3.125.3